Temporarily make Visitor, Editor non-abstract
If they’re not abstract, the compiler will tell me if I forgot to
give them their default implementation (visitX in Visitor, copy in
Editor), which is very useful during development. (This is how I found
the missing implementations in the previos commit.)

I’ll most likely revert this before the first release, as Visitor and
Editor are useless if you keep *all* the default implementations.