Add source_gen module to speed up my work
This will generate basic, “scaffolding” source code for a node type to
make my life easier.
**ceylon.ast source code will not be auto-generated on build.** This is
_not_ like the compiler’s ANTLR AST generation: I believe the ceylon.ast
API is more complicated (e. g. how child nodes are populated to the
super initializer), to the point where any source-gen that should be
able to fully generate the source code would be overly complex.
Instead, I will generate the source code for a class _once_, refine it
by hand, and then commit it regularly. This will speed up my work, but
no one else needs to have this or use it. (It follows that I won’t
officially support it either.)

Usage:

    ceylon run source_gen.ceylon.ast Type SuperType ParamType1 paramName1 ParamType2 paramName2 ...

and then enter the documentation of the node; when you’re finished,
close stdin with Ctrl+D and the module will generate the class for you.

More features (adding to Transformer, Editor, Visitor, CET, RHT, and
adding backend files) coming soon.

Thanks @FroMage for the idea! (On the ceylon-dev mailing list)