source-gen: Also make doc verbatim for double backticks
In 4aded81, I changed source-gen to make the doc string verbatim if it
contained quotes. I overlooked that double backticks can also disrupt a
string, requiring to make it verbatim as well.

(Also, the fix was completely missing for aliases and abstract classes
in the first place. Oops… I suppose copy+paste isn’t the right answer,
but I can’t be bothered with the proper solution right now.)