Remove unnecessary imports (same package)
Many of the backend conversion files had the following import:

    import ceylon.ast.redhat {
        createParser
    }

which is completely unnecessary, since they’re *in* the
ceylon.ast.redhat package. (I suspect it was originally added by the
IDE, and then I copied it around, not realizing it’s not needed, until
it ended up even in source-gen.)

Removed with the following command:

    grep -lr '^    createParser$' source/ceylon/ast/redhat/ | \
        while read file; do
            sed -i '/import ceylon.ast.redhat {/,+2d' "$file";
        done

and then I had to fix BaseType and QualifiedType manually because these
imported two things from the package, in which case the above command
left behind an extra closing brace.