RedHatTransformer: Fix token order in Bounds
Depending on whether a Bound is used as a lower or an upper bound,
its token (< / <=) appears before or after the endpoint, which affects
the correct token order. We now record the information whether it’s a
lower or upper bound in the node’s extra information (#17) when
transforming the WithinOperation, and then use that when transforming
the Bound itself.