Add test.ceylon.ast.samples module
Runs some “big picture”-style tests on the AST – here, converting +
formatting the sample hello world program and comparing it to expected
code.

The tests in this module aren’t unit tests, and they aren’t meant to
find new bugs; they’re rather intended for reassurance, to verify that
the AST really works (which I’m not entirely confident of in the RedHat
backend tests because these construct() everything, so there’s no
visible code to compare).

It’s also a nice usage sample of how to write the AST to code.