Add Variable, TypedVariable
No RedHat AST conversion, because it can’t be done on the variable node
level. For example, even though the RedHat Variable has a type field,
it’s not used for typed variables as they appear in “is” conditions,
where instead the type is stored in the condition itself.

(For the same reason, there’s no test either.)