Add tests for InitialComprehensionClause + subs
I’m working on my netbook at the moment, which isn’t really powerful
enough to run the Ceylon IDE with ceylon.ast, so this was all done in
the command line with emacs and ant… therefore, I don’t have the
patience required to add the dozen or so test cases that are necessary
to add meaningful comprehension test cases. So long story short, the
tests suck. Sorry about that.