Update ceylon.ast.samples module documentation
I’m also going to add Visitor and Editor samples. In a way, this
demonstrates the complete “life cycle” of an AST: creation, editing, and
comsumption. (The analogy is a bit far-fetched, though; you don’t
necessarily have to edit an AST, and of course you can continue to edit
+ visit it after the first “consumption”.)

Closes #51, I suppose, because ceylon.ast.samples now kinda takes on the
role of the ceylon.ast.util proposed there. It kinda sucks that users of
the “utility” components also import the mostly useless sample ASTs, but
on the other hand that’s probably true for a dedicated ceylon.ast.util
module – just because you need one of the utility visitors or editors
doesn’t mean that you’ll need all the others. (My suggested “solution”
would be that users just copy the visitor / editor they need.)