Add ValueArgument (RedHat’s AttributeArgument)
As it happens, while thinking about examples for the class
documentation, I realized that they’re actually really useful! The
RedHat AST → ceylon.ast conversion contains what is (as far as I recall)
the first use of value arguments in ceylon.ast… but I love it. I think
I’ll rewrite all the conversions like this (when I have the time). It’s
a lot cleaner this way.