Add sample package: bodyReturnRewrite
I wanted to figure out the algorithm that determines whether a body has
early returns (for ceylon/ceylon-compiler#1794 / ceylon/ceylon-compiler#1798),
but didn’t feel like fighting with the javac AST, so I started to write
it in pseudocode instead. I quickly realized that my pseudocode was
really mostly Ceylon, and that it’s actually not that bad of a
demonstration of ceylon.ast: It’s a real-world problem where both the
problem and the solution are easy enough to understand, and it
demonstrates the key concepts of Visitor and Editor.

(Of course, this won’t actually be used in the compiler… I’ll port it to
Java + javac AST.)

Editor still to be done.