Swap parameters of Parameter classes
Back when I wrote these, I wasn’t sure if I wanted to make the order of
the class parameters different than the order of the children in the
actual quote (i. e. put annotations last so they can be defaulted, even
though in code annotations are the first element). Later, in all other
declarations, I did that (annotations are always the last parameter,
even when there are other defaulted parameters as well), so I’m now
backporting that decision to Parameter as well.