Test type system Turing machine
Tests the type system Turing from the previous commit with a Turing
machine that tests if the length of the input is a power of two. The
test involves typechecking the compilation unit and then checking if we
got exactly the expected error on the function that tested bad input.

Thanks @quintesse for helping me with the typechecker usage! This was a
lot harder than expected…