Use more precise children types
These were long prevented by a compiler bug that would produce javac
errors like these:

    ImportTypeElement.ceylon:24: error: incompatible types
              children = [importAlias, name, nestedImports];
                                             ^
    required: Sequential<? extends Node>
    found:    Sequence<CAP#1>
    where CAP#1 is a fresh type-variable:
      CAP#1 extends Object from capture of ? extends Object

These have been reported multiple times (each with a “narrowed down”
version that was fixed, but oversimplified the ceylon.ast situation):

- ceylon/ceylon-compiler#1728
- ceylon/ceylon-compiler#1834
- ceylon/ceylon-compiler#2026

The last issue is still open; it is suspected that the bug which prevented
compiling this was fixed by ceylon/ceylon-compiler#2048, with the
underlying issue of ceylon/ceylon-compiler#2026 remaining unfixed.
However, that doesn’t matter to us: this code now compiles, so we can
finally merge it.