More missing native annotations
I’d forgotten that the CeylonExpressionTransformer test needs JVM-only
ceylon.file. I want the ceylon.ast.core tests to be backend-independent
though, so I made just the test itself native. (It’s not very good on
JS, of course.)

With ceylon/ceylon-skd#239, it should be possible to revert this commit.