Allow some more return types
Callable parameters and function declarations may now also write their
return type as `dynamic` or `function`. Part of this is in compliance
with ceylon/ceylon-spec#1303; the other part is that not allowing
`function` in these places is a semantic restriction, not a syntactical
one, and thus doesn’t belong in ceylon.ast.