Support length-based abbreviation for tuple types
A SequencedType can now optionally have an IntegerLiteral type to form a
tuple of the given length.

The syntax Boolean[8] has been around for a while (even before the
Ceylon 1.1 release), but was only recently declared an officially
supported feature of the language, part of the specification. See
ceylon/ceylon-spec#1041.