Fix missing updates for #87
Holy bejeezus this took a long time. But now all nodes should be updated
with *some* original node. In a lot of cases, the original node doesn’t
make a lot of sense though, so this is definitely up for debate. If you
want some of this handled differently, feel free to go yell at me or
something.