RedHatTransformer: sort methods alphabetically
Most of the disorder came from the PrecedenceXExpression rename from
1a45644; some also came from uncertainty as to how exactly mixed-case
names should be sorted.

Sorting was checked with the following command:

    grep '^    shared actual' source/ceylon/ast/redhat/RedHatTransformer.ceylon | \
        grep -o '[a-zA-Z]*(' | \
	sort -fc

This also clarifies how sorting is to be done: independent of case. In
the future, I hope to apply this to imports as well, and perhaps
retroactively fix unsorted imports.