Add Extension, Construction
These will replace ClassInstantiation, though I’ll do that in a separate
commit. The design is mostly taken from the spec, except that I inlined
four auxiliary nodes.

The design of the RedHat AST, on the other hand, is decidedly _not_
taken from the spec. There’s a lot of history in it, most importantly
the idea that constructors would be reified types. As a result, the
conversion is pretty painful.