Improve solution to #92
The previous fix didn’t work if the literal started and ended with
escaped quotes, since LiteralVisitor would unescape them, leading us to
think that this literal hadn’t been visited by LiteralVisitor. Instead,
we now compare the strings of the literal and the token – LiteralVisitor
sets the former, but not the latter, so if they differ, we know
LiteralVisitor was there. (It is of course possible that LiteralVisitor
set a text that didn’t differ from the old one, but in that case it
doesn’t hurt if we attempt to strip spaces either.)

Also, string template conversion now uses the same function as string
literal conversion. Previously, multi-line string templates weren’t
converted correctly.