Add missing RedHat AST infrastructure for types
The parse and compile functions for SimpleType, PrimaryType, MainType,
and UnionableType were missing, as well as tests for all of them plus
Type. Most likely a relic from the time before all of this
infrastructure was being auto-generated by source-gen.

Noticed in #52 and #100; fixes #100.

Note that tests currently don’t pass; this is expected since we have two
implementations of #52 right now, so the backwards conversion for one of
them must fail. This will be resolved in a separate commit which will
remove one of the implementations.