make Visitor not satisfy WideningTransformer
Visitors that satisfy WideningTransformer turn out to be measurably slower than
WideningTransformer's themselves. This commit makes Visitor independent of
WideningTransformer for a nice speed boost.

Steps to recreate the bulk of this patch:

1. Open the nodes in vi:
    vi `grep -l 'shared actual Result transform' * \
        | grep -v SelfReference`

2. Record a macro:
    - qq // start recording
    - /shared actual Result transform // find transform()
    - <shift-v>jjy // highlight & copy
    - jj // move down
    - p // paste
    - DAshared actual void visit(Visitor visitor)
    - <esc>j
    - :s/transformer/visitor
    - :s/transform/visit
    - jhhhD // whitespace
    - q // stop recording

3. Undo changes (uuuuuuuuuuuuu)

4. Move to top of file (:1)

5. Execute macro on all files:
    :bufdo execute "normal @q" | write

6. :q

7. Manually fix
    - SelfReference.ceylon
    - Identifier.ceylon

----

Committer’s note: the above notes apply to an OS X system. On a
GNU/Linux system, you may need to explicitly use vim instead of vi,
depending on how your distribution supplies the vi program (this needs
vim and will not work with the original vi).