Avoid unnecessary copy when editing leaf nodes
Suggested by @jvasileff in #111: for leaf nodes (that is, nodes without
any children), calling copy() is pointless, since there are no child
nodes that could be edited, so the returned node will be equal to the
original one. So in that case, we can skip the editing and keep the
original node. I don’t know if the copy itself is very expensive, but
this probably removes some GC pressure, at least.

Done with the following command (requires GNU sed):

    sed -i 's/=> that.copy();/=> that;/' source/ceylon/ast/core/Editor.ceylon

Documentation updated manually.

A test which checked that Editor actually created new nodes was removed,
since this new behavior breaks it.