Fix c.a.samples.turingMachine.powerOfTwo test
The typechecker now adds a “with strict null checking” to the error
message before the colon. I don’t want to rely on that, so let’s just
not require that the 'accept' name is immediately followed by a colon,
leaving some space for the “with strict null checking” as well as other
modifiers.