Update CaseClause.hash for elseCase flag (#137)
Updating node types is tricky – there are lots of places that need to be
updated, and many of them don’t cause any errors if you forget. I’ve
created the wiki page “changing nodes” to list all these places so that
this hopefully doesn’t happen again in the future.