Add help to source-gen helper scripts
Since I so rarely use them, I usually have to look into the source-gen
run.ceylon to remember the arguments. No more!

(The Ceylon code does have a --help option, but the abstract and alias
scripts hide it from you, and it also only points you to an ancient and
outdated gist that’s not really helpful at all.)