Add support for constant definitions in module body
Part of #136. (Note that this doesn’t yet add support for *using* these
constants in a module *import*.)

The assertion in ModuleBody isn’t very pretty, and I’m also not sure if
ValueDefinition is the right type to use (the other option would be
AnyValue – the parser supports getters but the typechecker rejects
them). This will hopefully be cleaned up somewhat once the feature has a
specification behind it.