Fix conversion of prefixed identifiers
The parser used to set the text of prefixed identifiers’ tokens to the
identifier name (without the prefix), and we could only distinguish them
from unprefixed identifiers via the token’s length. With
eclipse/ceylon@4c45a73508, this was changed so that the token’s text is
untouched, and the prefix is stripped away later. This means we detect
prefixed tokens differently now and also have to produce different
tokens in the RedHatTransformer. It also means we can remove the
optional length argument of TokenFactory.token(), because its only use
is now obsolete.

See also eclipse/ceylon.formatter#144 for related formatter changes.