Add support for new destructuring syntax
A Destructure is now a simple wrapper around a PatternList, similar to a
LetExpression (but without the accompanying expression). The old syntax
is supported when parsing or converting from the compiler AST to
ceylon.ast, but not as a separate syntax in ceylon.ast.core. (In fact,
we get that support mostly for free because both syntaxes parse to
equivalent ASTs anyways.) Fixes #138.