Rearrange member operators in complete compilation unit
It seems that since 1.3.3, the compiler’s parser lost the ability to
parse `package?.something` and `package*.something` as qualified
expressions (possibly in eclipse/ceylon@5c80f1de3e). We already use a
SpreadMemberOperator elsewhere, so we can simply drop it from the
package self reference. And while we’re at it, also move the
SafeMemberOperator from the outer self reference to a slightly more
plausible location (sysout could at least in theory be null).