Switch parsing order in parseAnySpecifier
Since eclipse/ceylon#7190, an expression like `=>a*b+c` is valid: it is
a function expression with zero parameter lists (see also #142). And
apparently the parser will parse it as such when asked to parse this
code as a (non-lazy) specifier (though I’m not sure why), whereas we
actually want this to be parsed as a lazy specifier in this context. We
can avoid this by attempting the lazy specifier parse first and falling
back to the non-lazy case second. (Even better would be to only call the
parser once, but there’s no direct rule for “any specifier”, and I’m not
convinced that the alternative – turning the code into a complete
example for some other rule, e. g. typedMethodOrAttributeDeclaration –
would be better.)